refactor group joiners
